+++ ui/src/stories/RadioButtons.stories.tsx
@@ -35,3 +35,5 @@ radioButtons.args = {
   items: radios,
   onChange: () => {},
 };
+
+radioButtons.storyName = 'RadioButtons';
